refactor: class 형 함수로 수정
