--- application/usecase/musics/dto/Video.dto.ts
@@ -1,7 +1,0 @@
-export interface Video {
-  title: string;
-  channelTitle: string;
-  thumbnail: string;
-  duration: string;
-  musicList: string[];
-}
